+++ apps/backend/apps/backend/medusa-config.ts
@@ -57,5 +57,8 @@ module.exports = defineConfig({
         ],
       },
     },
+    {
+      resolve: "./src/modules/seo",
+    },
   ],
 })